PCQ-515: Update to remove iterating services during ES.
