+++ src/functionalTest/resources/JsonTestFiles/FirstSubmitAnswer.json
@@ -4,7 +4,7 @@
   "partyId": "test@test2.com",
   "channel": 1,
   "completedDate": "2020-03-05T09:13:45.000Z",
-  "serviceId": "PROBATE_TEST",
+  "serviceId": "PCQTEST",
   "actor": "DEFENDANT",
   "versionNo": 1,
   "pcqAnswers": {
